fix: Move the menu items into the correct place, and deprecate the old window.
