+++ Assets/Plugins/Source/Editor/EditorWindows/EOSSettingsWindow.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bf93ccfa48b65594caf161002f4d3562
+guid: 43e463f37d0d1024eb3c8cff22f5f0be
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
